Update cql to elm tests
